refactor(ui): move the grid skeleton next to poster_card_body
The placeholder imitates poster_card_body, so it belongs beside it. Its
wrapper now carries the same shadow-lg and h-full as the landed
trending_card chrome, minus the interaction-only classes a placeholder
does not need.

Re-measured height parity in a real browser after the min-h-[2lh] title
fix landed: a real card now measures the same height regardless of
whether its title wraps to one line or two, and the skeleton already
matches it exactly in both themes, with no further tuning needed.
